smallestUniqueGroups {msqrob2}	R Documentation

Smallest unique protein groups

Description

For a given vector of protein group names, outputs the names of those protein groups for which none of its member proteins is present in a smaller protein group.

Usage

smallestUniqueGroups(proteins, split = ";")

Arguments

proteins
A vector of characters or factors containing single proteins and/or protein groups (i.e. proteins separated by a separator symbol).

split
The character string that is used to separate the indivudual protein names in each protein group.

Value

A character vector containing the names of the protein groups for which none of its proteins is present in a smaller protein group.

Examples

data(pe)
smallestUniqueGroups(rowData(pe[["peptide"]])$Proteins)

[Package msqrob2 version 1.2.0 Index]