getModel,StatModel-method {msqrob2}	R Documentation

Accessor functions for StatModel class

Description

Accessor functions for StatModel class

getModel(object)

to get model

getFitMethod(object)

to get the parameter estimation method

getCoef(object)

to get the parameter estimates of the mean model

getDF(object)

to get the residual degrees of freedom of the model

getVar(object)

to get the residual variance of the model

getSigma(object)

to get the residual standard deviation of the model

getDfPosterior(object)

to get the degrees of freedom of the empirical Bayes variance estimator

getVarPosterior(object)

to get the empirical Bayes variance

getSigmaPosterior(object)

to get the empirical Bayes standard deviation

getVcovUnscaled(object)

to get the unscaled variance covariance matrix of the model parameters

Usage

## S4 method for signature 'StatModel'
getModel(object)

## S4 method for signature 'StatModel'
getFitMethod(object)

## S4 method for signature 'StatModel'
getCoef(object)

## S4 method for signature 'StatModel'
getDfPosterior(object)

## S4 method for signature 'StatModel'
getVarPosterior(object)

## S4 method for signature 'StatModel'
getSigmaPosterior(object)

## S4 method for signature 'StatModel'
getDF(object)

## S4 method for signature 'StatModel'
getVar(object)

## S4 method for signature 'StatModel'
getSigma(object)

## S4 method for signature 'StatModel'
getVcovUnscaled(object)

Arguments

object
StatModel object

Value

The requested parameter of the StatModel object

Examples

data(pe)

# Aggregate peptide intensities in protein expression values
pe <- aggregateFeatures(pe, i = "peptide", fcol = "Proteins", name = "protein")

# Fit msqrob model
pe <- msqrob(pe, i = "protein", formula = ~condition)
getCoef(rowData(pe[["protein"]])$msqrobModels[[1]])
getModel(rowData(pe[["protein"]])$msqrobModels[[1]])
getFitMethod(rowData(pe[["protein"]])$msqrobModels[[1]])
# Similar for the remaining accessors

[Package msqrob2 version 1.2.0 Index]